Moved I18n Annotation => org.vaadin.spring.i18n.annotation
